requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.100.0
youtube-search-python==1.6.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
tweepy==4.14.0
instaloader==4.10.3
pandas==2.1.3
numpy==1.26.2
textblob==0.17.1
vaderSentiment==3.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
openpyxl==3.1.2

